DiagnosticsFeature + DiagnosticsServer working
