+++ Jasper.sln
@@ -1,4 +1,4 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
@@ -16,6 +16,8 @@ EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "StorytellerSpecs", "src\StorytellerSpecs\StorytellerSpecs.xproj", "{AD8B964A-588D-4376-9522-21E97BBA4335}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "TestMessages", "src\TestMessages\TestMessages.xproj", "{80F9678C-6935-4048-AACF-41642A48282A}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Jasper.Diagnostics", "src\Jasper.Diagnostics\Jasper.Diagnostics.xproj", "{717D1E6D-015E-4924-89C4-81C1E277A62F}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "JasperHttp", "src\JasperHttp\JasperHttp.xproj", "{D4A3B31B-67EC-41C3-A514-CD75FAD1E573}"
 EndProject
@@ -57,6 +59,10 @@ Global
 		{80F9678C-6935-4048-AACF-41642A48282A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{80F9678C-6935-4048-AACF-41642A48282A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{80F9678C-6935-4048-AACF-41642A48282A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{717D1E6D-015E-4924-89C4-81C1E277A62F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{717D1E6D-015E-4924-89C4-81C1E277A62F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{717D1E6D-015E-4924-89C4-81C1E277A62F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{717D1E6D-015E-4924-89C4-81C1E277A62F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D4A3B31B-67EC-41C3-A514-CD75FAD1E573}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D4A3B31B-67EC-41C3-A514-CD75FAD1E573}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D4A3B31B-67EC-41C3-A514-CD75FAD1E573}.Release|Any CPU.ActiveCfg = Release|Any CPU
