+++ global.json
@@ -1,6 +1,6 @@
 {
   "projects": [ "src", "test" ],
   "sdk": {
-  "version": "1.0.0-preview2-003131"
+  "version": "1.0.0-preview2-1-003177"
   }
 }
