+++ src/DiagnosticsHarnessMessages/Messages.cs
@@ -1,0 +1,11 @@
+﻿namespace DiagnosticsHarnessMessages
+{
+    public class MiddlewareMessage
+    {
+        public string Message { get; set; }
+    }
+
+    public class AMessageThatWillError
+    {
+    }
+}
