+++ src/DiagnosticsHarnessMessages/project.json
@@ -1,0 +1,20 @@
+{
+  "version": "1.0.0-*",
+
+  "dependencies": {
+  },
+
+  "frameworks": {
+    "netstandard1.5": {
+      "dependencies": {
+        "NETStandard.Library": "1.6.1"
+      },
+      "imports": [ "dnxcore50" ]
+    }
+  },
+
+  "buildOptions": {
+    "debugType": "portable"
+  }
+
+}
