+++ src/DiagnosticsServiceHarness/project.json
@@ -1,19 +1,18 @@
-﻿{
+{
+  "version": "1.0.0-*",
+
   "dependencies": {
     "Jasper.Diagnostics": { "target": "project" },
     "JasperBus": { "target": "project" },
     "DiagnosticsHarnessMessages": { "target": "project" },
     "StructureMap": "4.4.3",
     "StructureMap.Microsoft.DependencyInjection": "1.3.0",
+    "PeterKottas.DotNetCore.WindowsService": "1.0.16",
     "Microsoft.NETCore.App": {
       "version": "1.1.0",
       "type": "platform"
     },
-    "Microsoft.AspNetCore.Mvc": "1.1.0",
-    "Microsoft.AspNetCore.Server.IISIntegration": "1.1.0",
-    "Microsoft.AspNetCore.Server.Kestrel": "1.1.0",
     "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.1.0",
-    "Microsoft.Extensions.Configuration.FileExtensions": "1.1.0",
     "Microsoft.Extensions.Configuration.Json": "1.1.0",
     "Microsoft.Extensions.Logging": "1.1.0",
     "Microsoft.Extensions.Logging.Console": "1.1.0",
@@ -38,26 +37,6 @@
   "buildOptions": {
     "emitEntryPoint": true,
     "preserveCompilationContext": true
-  },
-
-  "runtimeOptions": {
-    "configProperties": {
-      "System.GC.Server": true
-    }
-  },
-
-  "publishOptions": {
-    "include": [
-      "stats.json",
-      "wwwroot",
-      "Views",
-      "Areas/**/Views",
-      "appsettings.json",
-      "web.config"
-    ]
-  },
-
-  "scripts": {
-    "postpublish": [ "dotnet publish-iis --publish-folder %publish:OutputPath% --framework %publish:FullTargetFramework%" ]
   }
+
 }
