+++ src/Jasper.Diagnostics/app/Live/SentMessages.js
@@ -1,0 +1,14 @@
+import {
+  default as React,
+  PropTypes
+} from 'react'
+import { connect } from 'react-redux'
+import Envelope from '../Components/Envelope'
+
+const SentMessages = () => {
+  return (
+    <div>Sent Messages</div>
+  )
+}
+
+export default SentMessages
