+++ src/Jasper.Diagnostics/project.json
@@ -34,12 +34,13 @@
     "JasperBus": { "target": "project" },
     "Newtonsoft.Json": "9.0.1",
     "StructureMap": "4.4.3",
+    "StructureMap.AspNetCore": "1.0.0",
     "StructureMap.Microsoft.DependencyInjection": "1.3.0",
-    "Microsoft.AspNetCore.Hosting": "1.0.0",
+    "Microsoft.AspNetCore.Hosting": "1.1.0",
     "Microsoft.AspNetCore.Http.Abstractions": "1.1.0",
     "Microsoft.AspNetCore.Http.Extensions": "1.1.0",
-    "Microsoft.AspNetCore.Server.Kestrel": "1.0.0",
-    "Microsoft.AspNetCore.StaticFiles": "1.0.0",
+    "Microsoft.AspNetCore.Server.Kestrel": "1.1.0",
+    "Microsoft.AspNetCore.StaticFiles": "1.1.0",
     "Microsoft.AspNetCore.WebSockets": "1.0.0"
   },
   "frameworks": {
